refactor(api): modularizar enpoints y modelos de datos de notificaciones y canales rss
