+++ newsradar_api/app/schemas/rss.py
@@ -1,0 +1,21 @@
+from typing import Optional
+from pydantic import BaseModel, Field, HttpUrl
+
+
+class RSSChannelBase(BaseModel):
+    url: HttpUrl
+    category_id: int
+
+
+class RSSChannelCreate(RSSChannelBase):
+    pass
+
+
+class RSSChannelUpdate(BaseModel):
+    url: Optional[HttpUrl] = None
+    category_id: Optional[int] = None
+
+
+class RSSChannel(RSSChannelBase):
+    id: int
+    information_source_id: int
